+++ hv_almalinux84.ps1
@@ -4,9 +4,9 @@
 $startDTM = (Get-Date)
 
 # Variables
-$template_file="./templates/hv_almalinux8_g2.json"
-$var_file="./variables/variables_almalinux83.json"
-$machine="AlmaLinux 8.3"
+$template_file="./templates/hv_almalinux8_g2.pkr.hcl"
+$var_file="./variables/variables_almalinux84.pkvars.hcl"
+$machine="AlmaLinux 8.4"
 $packer_log=0
 
 if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {