+++ hv_almalinux84_vagrant.ps1
@@ -4,10 +4,9 @@
 $startDTM = (Get-Date)
 
 # Variables
-$template_file="./templates/hv_almalinux8_g2_docker.json"
-$var_file="./variables/variables_almalinux83.json"
-$override="./variables/almalinux8_docker.yml"
-$machine="AlmaLinux 8.3"
+$template_file="./templates/hv_almalinux8_g2_vagrant.pkr.hcl"
+$var_file="./variables/variables_almalinux84.pkvars.hcl"
+$machine="AlmaLinux 8.4"
 $packer_log=0
 
 if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {
@@ -15,7 +14,7 @@ if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {
   Write-Output "Building: $machine"
   try {
     $env:PACKER_LOG=$packer_log
-    packer validate -var-file="$var_file" -var "ansible_override=$override" "$template_file"
+    packer validate -var-file="$var_file" "$template_file"
   }
   catch {
     Write-Output "Packer validation failed, exiting."
@@ -24,7 +23,7 @@ if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {
   try {
     $env:PACKER_LOG=$packer_log
     packer version
-    packer build --force -var-file="$var_file" -var "ansible_override=$override" "$template_file"
+    packer build --force -var-file="$var_file" "$template_file"
   }
   catch {
     Write-Output "Packer build failed, exiting."